refactor(set): add validation and deleted-state checks

+++ src/main/kotlin/com/sentencefit/application/teacherset/service/DeleteSetService.kt
@@ -23,6 +23,10 @@ class DeleteSetService(
         val studySet = loadSetPort.findByTeacherIdAndClassIdAndId(teacherId, classId, setId)
             ?: throw SetException(SetErrorCode.SET_NOT_FOUND)
 
+        if (studySet.status.name == "DELETED") {
+            throw SetException(SetErrorCode.SET_ALREADY_DELETED)
+        }
+
         saveSetPort.save(studySet.delete())
     }
 }
